requirements.txt
python-dotenv
pymongo
pandas 
numpy
certifi
-e .
scikit-learn
scipy
dill
pyaml
mlflow
dagshub